% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos_Escorrentia.R
\name{Datos_Escorrentia}
\alias{Datos_Escorrentia}
\title{Escorrentía}
\usage{
Datos_Escorrentia(Precipitacion, precipitacion_efectiva)
}
\arguments{
\item{Precipitacion}{Raster stack de Precipitación.}

\item{Precipitacion_efectiva}{Raster stack de precipitación efectiva.}
}
\value{
Devuelve un raster stack con datos de escorrentia  limitado a la zona de estudio.
}
\description{
Realiza la estimación de la esconrrentía superficial
}
\details{
La ecuación utilizada representa la diferencia entre la precipitación y la precipitación efectiva.
}
