% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos_Precipitacion.R
\name{Datos_Precipitacion}
\alias{Datos_Precipitacion}
\title{Precipitación}
\usage{
Datos_Precipitacion(Area)
}
\arguments{
\item{Area}{Vectorial de la zona de estudio con la proyección EPSG: 4326.}
}
\value{
Devuelve un raster stack con datos de precipitación limitado a la zona de estudio.
}
\description{
Descarga datos de precipitación de worldclim.org y delimitado a la zona de estudio.
}
\details{
Descarga los datos de precipitación
}
