% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datos_Precipitacion_efectiva.R
\name{Datos_Precipitacion_efectiva}
\alias{Datos_Precipitacion_efectiva}
\title{Precipitación Efectiva}
\usage{
Datos_Precipitacion_efectiva(Precipitacion)
}
\arguments{
\item{Precipitacion}{Raster stack de Precipitación.}
}
\value{
Devuelve un raster stack con datos de precipitación  limitado a la zona de estudio.
}
\description{
Realiza la estimación de precipitación efectiva
}
\details{
La ecuación utilizada es obtenida del software Cropwat, compara el valor de precipitación y aplica la ecuación correspondiente.
Si P<251... PE= P(PE x 125-0.2 x P)/125
Si => 251... PE= 125+0.1 x P
}
