% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descarga_MODIS.R
\name{Descarga_MODIS}
\alias{Descarga_MODIS}
\title{Descarga de MOD16A2 (Evapotranspiración).}
\usage{
Descarga_MODIS(Zona)
}
\arguments{
\item{Zona}{Es el área de estudio cargado con la función Zona_estudio.}
}
\description{
Descarga datos de evapotranspiración
}
\details{
La función utiliza como parámetro de entrada un vectorial para delimitar los datos a descargar dentro de un rango de fechas que corresponden al ciclo vegetativo del cultivo.
}
