% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Escorrentia.R
\name{Escorrentia}
\alias{Escorrentia}
\title{Escorrentía}
\usage{
Escorrentia()
}
\value{
Devuelve un raster stack con los datos de escorrentía generada.
}
\description{
Estima la escorrentía generada.
}
\details{
Carga datos geoespaciales de precipitación y precipitación efectiva anteriormente generado en las funciones de Precipitación y Precipitación_efectiva.
}
