% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Global_RR.R
\name{Global_RR}
\alias{Global_RR}
\title{Requerimiento hidrico}
\usage{
Global_RR()
}
\description{
Realiza la estimación del requerimiento de riego
}
\details{
Realiza la estimación del requerimiento de riego, seleccionando área de estudio, descarga y procesado de datos MOD16A2, descarga y procesado de precipitación, estimación de la precipitación efectiva y estimación del requerimiento de riego.
Genera un reporte y gráfico de Evapontraspiración de referencia, Precipitación efectiva y requerimiento de riego.
}
