% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_datos.R
\name{Lectura_MODIS}
\alias{Lectura_MODIS}
\title{Datos MODIS}
\usage{
Lectura_MODIS(Zona)
}
\arguments{
\item{Zona}{Es el vectorial cargado en la función Zona_estudo.}
}
\value{
Devuelve un raster stack de los datos MOD16A2.
}
\description{
Lectura de datos MOD16A2 y procesamiento.
}
\details{
Carga los datos MOD16A2 y realiza el procesamiento  para el área de estudio.
}
