% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIS.R
\name{MODIS}
\alias{MODIS}
\title{Datos MODIS de Evapotranspiración}
\usage{
MODIS(Area)
}
\arguments{
\item{Area}{Vectorial de la zona de estudio con la proyección EPSG: 4326.}
}
\value{
Devuelve un raster stack de imagenes de evapotranspiración.
}
\description{
Descarga el producto MOD16A2, procesa las imagenes a una zona de estudio.
}
\details{
Inicia solicitando el rango de fechas, verificando que las fechas sean correctas.
Realiza una validación que se encuentre instalado OSGEO para el uso de las librerias OGR/GDAL.
Solicita usuario y contraseña USGS para la descarga de imagenes MOD16A2.
Descarga y realiza el procesamiento de MOD16A2.
Realiza una interpolación para los datos faltantes.
Finaliza con la creación de imagenes y archivos raster.
}
