% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modis_mes.R
\name{Modis_mes}
\alias{Modis_mes}
\title{Composición mensual}
\usage{
Modis_mes(Area, Modis_datos)
}
\arguments{
\item{Area}{Vectorial de la zona de estudio.}

\item{Modis_datos}{Raster stack de Evapotranspiración.}
}
\value{
Devuelve un raster stack con datos de requerimiento de riego limitado a la zona de estudio.
}
\description{
Suma los datos acorde al mes para obtener Evapotranspiración mensual
}
