% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitacion.R
\name{Precipitacion}
\alias{Precipitacion}
\title{Datos de precipitaci贸n}
\usage{
Precipitacion(Zona)
}
\arguments{
\item{Zona}{Es el archivo vectorial cargado anteriormente con la funci贸n ZOna_estudio}
}
\value{
Devuelve un raster stack de datos de precipitaci贸n.
}
\description{
Descarga datos de worldclim.org.
}
\details{
Descarga datos geoespaciales de precipitaci贸n del portal worldclim.org, posteriormente es procesado a la zona de estudio.
}
