% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Precipitacion_efectiva.R
\name{Precipitacion_efectiva}
\alias{Precipitacion_efectiva}
\title{Precipitación efectiva}
\usage{
Precipitacion_efectiva()
}
\value{
Devuelve un raster stack de datos de precipitación efectiva.
}
\description{
Estima la precipitación efectiva.
}
\details{
Lee los archivos de precipitación creados anteriormente con la función Precipitacion.
}
