% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Req_Hid.R
\name{Req_Hid}
\alias{Req_Hid}
\title{Estimación de requerimiento de riego}
\usage{
Req_Hid()
}
\description{
Realiza la descarga los datos y el procesamiento de las imágenes para obtener el requerimiento hídrico de un cultivo basado en el rango temporal del ciclo vegetativo.
}
