% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Requerimiento.R
\name{Requerimiento}
\alias{Requerimiento}
\title{Requerimiento de riego}
\usage{
Requerimiento(Zona)
}
\arguments{
\item{Zona}{Es el archivo vectorial cargado anteriormente con la funci贸n ZOna_estudio.}
}
\value{
Devuelve un raster stack de requerimiento de riego.
}
\description{
Estima el requerimiento de riego con los datos de evapotranspiraci贸n de MODIS y Precipitaci贸n de worldclim.org.
}
\details{
Descarga datos geoespaciales de precipitaci贸n del portal worldclim.org, posteriormente es procesado a la zona de estudio.
}
