% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Volumen.R
\name{Volumen}
\alias{Volumen}
\title{Volumen}
\usage{
Volumen(mm, Area)

Volumen(mm, Area)
}
\arguments{
\item{mm}{Raster con valores en mm de evapotranspiración.}

\item{Area}{Vectorial de la zona de estudio.}
}
\value{
Devuelve un raster stack con datos de requerimiento de riego limitado a la zona de estudio.

Devuelve un raster stack con datos de requerimiento de riego limitado a la zona de estudio.
}
\description{
Realiza la estimación del volumen

Realiza la estimación del volumen
}
\details{
Estima el volumen en un área determinada.

Estima el volumen en un área determinada.
}
