% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zona_estudio.R
\name{Zona_estudio}
\alias{Zona_estudio}
\title{Delimitar a la zona de estudio.}
\usage{
Zona_estudio()
}
\value{
Devuelve la zona vectorial.
}
\description{
Carga un vectorial y devuelve un vectorial para ser usado en otras funciones.
}
\details{
Esta funci칩n es requerida para ser usada en otras funciones, debido a que limita a la zona de estudio los datos geoespaciales.
}
\concept{**Zona estudio** se refiere al 치rea geogr치fica de donde se realizar치 el proceso geoespacial.}
\keyword{Area}
\keyword{Vectorial}
