% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_mes.R
\name{Modis_mes}
\alias{Modis_mes}
\title{Composición mensual de Evapotranspiración}
\usage{
Modis_mes(Zona)
}
\arguments{
\item{Zona}{Vectorial cargado de la función Zona_estudio().}
}
\value{
Devuelve un raster stack de evapotranspiración mensual.
}
\description{
Crea composiciones mensuales de los datos geoespaciales.
}
\details{
Carga los datos de evapotranspiración generados con la función de ET_datos. Crea composiciones raster de 1 mes acorde al rango de fecha de los productos procesados.
}
