% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Requerimiento.R
\name{Requerimiento}
\alias{Requerimiento}
\title{Requerimiento de riego}
\usage{
Requerimiento(ET, PE, Area)
}
\arguments{
\item{ET}{Raster stack de Evapotranspiración.}

\item{PE}{Raster stack de precipitación efectiva.}

\item{Area}{Vectorial de la zona de estudio.}
}
\value{
Devuelve un raster stack con datos de requerimiento de riego limitado a la zona de estudio.
}
\description{
Realiza la estimación del requerimiento de riego
}
\details{
Realiza una homogenización de la resolución espacial entre los datos de entrada.
Solicita un archivo en Excel con valores de KC, en caso de no disponerlo puede ingresar un valor máximo, o dejarlo solo con la evapotranspiración de referencia para el cálculo.
Genera un reporte y gráfico de Evapontraspiración de referencia, Precipitación efectiva y requerimiento de riego.
}
