% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Zona_estudio.R
\name{Zona_estudio}
\alias{Zona_estudio}
\title{Delimitardor a zona de estudio}
\usage{
Zona_estudio()
}
\value{
Devuelve la zona de estudio cargada en R.
}
\description{
Lee un archivo vectorial de la zona de estudio, verifica que la proyecci칩n de coordendas sean geogr치ficas (EPSG:4326).
}
\details{
Lee un archivo vectorial, verifica el sistema de proyecci칩n de coordenadas en EPSG:4626, en caso de no coincidir se realiza la proyecci칩n correspondiente.
}
